extract PronounTrie to trie.rs
Again, this is just to keep the files short and easy to refactor.
